use float64 everywhere to fix inaccuracies caused by converting float32 <=> float64

+++ client/dev/osc/main.go
@@ -246,10 +246,10 @@ func loopPattern(pattern string) *osc.Bundle {
 	return bundle
 }
 
-func tempoMsg(offset int, bpm float32) *osc.Message {
+func tempoMsg(offset int, bpm float64) *osc.Message {
 	msg := osc.NewMessage("/system/tempo")
 	msg.Append(int32(offset))
-	msg.Append(float32(bpm))
+	msg.Append(bpm)
 	return msg
 }
